made the requested changes in the pull request
