+++ Services/AI-Register/AI-Register/AI-Register/Controllers/AISystemStatusController.cs
@@ -2,6 +2,7 @@
 using BusinessLogic.Classes;
 using BusinessLogic.Helpers;
 using Microsoft.AspNetCore.Mvc;
+using BusinessLogic.Enums;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -16,11 +17,11 @@ namespace AIRegister.Controllers
         public IActionResult Get()
         {
             List<EnumDTO> aiSystemStatusEnumCollection = new List<EnumDTO>();
-            foreach (string status in EnumHelper.EnumToList<AIRegisterEnum.AISystemStatus>())
+            foreach (string status in EnumHelper.EnumToList<AISystemStatus>())
             {
                 EnumDTO aiSystemStatusEnum = new EnumDTO();
                 aiSystemStatusEnum.Name = status;
-                aiSystemStatusEnum.id = EnumHelper.EnumParse<AIRegisterEnum.AISystemStatus>(status);
+                aiSystemStatusEnum.id = EnumHelper.EnumParse<AISystemStatus>(status);
                 aiSystemStatusEnumCollection.Add(aiSystemStatusEnum);
 
             }
@@ -32,7 +33,7 @@ namespace AIRegister.Controllers
         [HttpGet("{id}")]
         public IActionResult Get(int id)
         {
-            return Ok(EnumHelper.EnumToString<AIRegisterEnum.AISystemStatus>(id));
+            return Ok(EnumHelper.EnumToString<AISystemStatus>(id));
         }
 
     }
