+++ Services/AI-Register/AI-Register/AI-Register/DTOs/AIDetailDTO.cs
@@ -1,4 +1,5 @@
 ﻿using BusinessLogic.Classes;
+using BusinessLogic.Enums;
 
 namespace AIRegister.DTOs
 {
@@ -6,18 +7,18 @@ namespace AIRegister.DTOs
     {
         public Guid Guid { get; set; }
         public string Name { get; set; }
-        public AIRegisterEnum.AISystemStatus Status { get; set; }
+        public AISystemStatus Status { get; set; }
         public string Url { get; set; }
         public string TechnicalDocumentationLink { get; set; }
-        public AIRegisterEnum.ApprovalStatus ApprovalStatus { get; set; }
+        public ApprovalStatus ApprovalStatus { get; set; }
         public DateOnly DateAdded { get; set; }
         public string Description { get; set; }
         public ProviderDTO Provider { get; set; }
         public CertificateDTO Certificate { get; set; }
         public List<AISystemFile> Files { get; set; }
-        public AIRegisterEnum.MemberStates MemberState { get; set; }
+        public MemberStates MemberState { get; set; }
 
-        public AIDetailDTO(Guid guid, string name, AIRegisterEnum.AISystemStatus status, string url, string technicalDocumentationLink, AIRegisterEnum.ApprovalStatus approvalStatus, DateOnly dateAdded, ProviderDTO provider, CertificateDTO certificate, List<AISystemFile> files, string description, AIRegisterEnum.MemberStates memberState)
+        public AIDetailDTO(Guid guid, string name, AISystemStatus status, string url, string technicalDocumentationLink, ApprovalStatus approvalStatus, DateOnly dateAdded, ProviderDTO provider, CertificateDTO certificate, List<AISystemFile> files, string description, MemberStates memberState)
         {
             Guid = guid;
             Name = name;
