+++ Services/AI-Register/AI-Register/AI-Register/DTOs/GetAISystemDTO.cs
@@ -1,4 +1,5 @@
 ﻿using BusinessLogic.Classes;
+using BusinessLogic.Enums;
 
 namespace AIRegister.DTOs
 {
@@ -9,9 +10,9 @@ namespace AIRegister.DTOs
         public string ProviderName { get; set; }
         public DateOnly DateAdded { get; set; }
         public string Description { get; set; }
-        public AIRegisterEnum.ApprovalStatus ApprovalStatus { get; set; }
+        public ApprovalStatus ApprovalStatus { get; set; }
 
-        public GetAISystemDTO(Guid id , string name, string providerName, DateOnly dateAdded, AIRegisterEnum.ApprovalStatus approvalStatus, string description)
+        public GetAISystemDTO(Guid id , string name, string providerName, DateOnly dateAdded, ApprovalStatus approvalStatus, string description)
         {
             Id = id;
             Name = name;
