--- Services/AI-Register/AI-Register/Business Logic/Classes/AIRegisterEnum.cs
@@ -1,57 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BusinessLogic.Classes
-{
-    public class AIRegisterEnum
-    {
-        public enum ApprovalStatus
-        {
-            Approved = 1,
-            Pending = 2,
-            Rejected = 3
-        }
-
-        public enum AISystemStatus
-        {
-            OnMarket = 1,
-            InService = 2,
-            Recalled = 3
-        }
-
-        [Flags]
-        public enum MemberStates
-        {
-            Austria = 1 << 0,
-            Belgium = 1 << 1,
-            Bulgaria = 1 << 2,
-            Croatia = 1 << 3,
-            Cyprus = 1 << 4,
-            CzechRepublic = 1 << 5,
-            Denmark = 1 << 6,
-            Estonia = 1 << 7,
-            Finland = 1 << 8,
-            France = 1 << 9,
-            Germany = 1 << 10,
-            Greece = 1 << 11,
-            Hungary = 1 << 12,
-            Ireland = 1 << 13,
-            Italy = 1 << 14,
-            Latvia = 1 << 15,
-            Lithuania = 1 << 16,
-            Luxembourg = 1 << 17,
-            Malta = 1 << 18,
-            Netherlands = 1 << 19,
-            Poland = 1 << 20,
-            Portugal = 1 << 21,
-            Romania = 1 << 22,
-            Slovakia = 1 << 23,
-            Slovenia = 1 << 24,
-            Spain = 1 << 25,
-            Sweden = 1 << 26
-        }
-    }
-}
