+++ Services/AI-Register/AI-Register/Business Logic/Classes/AISystem.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using BusinessLogic.Entities;
+using BusinessLogic.Enums;
 using BusinessLogic.Interfaces;
 
 namespace BusinessLogic.Classes
@@ -12,15 +13,15 @@ namespace BusinessLogic.Classes
     {
         public Guid Guid { get; set; }
         public string Name { get; set; }
-        public AIRegisterEnum.AISystemStatus Status { get; set; }
+        public AISystemStatus Status { get; set; }
         public string URL { get; set; }
         public string Description { get; set; }
         public string TechnicalDocumentationLink { get; set; }
-        public AIRegisterEnum.ApprovalStatus ApprovalStatus { get; set; }
+        public ApprovalStatus ApprovalStatus { get; set; }
         public DateOnly DateAdded { get; set; }
         public Provider provider { get; set; }
         public Certificate certificate { get; set; }
-        public AIRegisterEnum.MemberStates MemberState { get; set; }
+        public MemberStates MemberState { get; set; }
         public ICollection<AISystemFile> Files { get; set; }
 
 
@@ -28,7 +29,7 @@ namespace BusinessLogic.Classes
         {
             
         }
-        public AISystem(Guid guid, string name, AIRegisterEnum.AISystemStatus status, string description, string url, string technicalDocumentationLink, AIRegisterEnum.ApprovalStatus approvalStatus, DateOnly dateAdded, Provider provider, Certificate certificate, AIRegisterEnum.MemberStates memberstate)
+        public AISystem(Guid guid, string name, AISystemStatus status, string description, string url, string technicalDocumentationLink, ApprovalStatus approvalStatus, DateOnly dateAdded, Provider provider, Certificate certificate, MemberStates memberstate)
         {
             Guid = guid;
             Name = name;
