+++ Services/AI-Register/AI-Register/Business Logic/Enums/AISystemStatus.cs
@@ -1,0 +1,8 @@
+﻿namespace BusinessLogic.Enums;
+
+public enum AISystemStatus
+{
+    OnMarket = 1,
+    InService = 2,
+    Recalled = 3
+}
