+++ Services/AI-Register/AI-Register/Business Logic/Enums/AprprovalStatus.cs
@@ -1,0 +1,8 @@
+﻿namespace BusinessLogic.Enums;
+
+public enum ApprovalStatus
+{
+    Approved = 1,
+    Pending = 2,
+    Rejected = 3
+}
