+++ Services/AI-Register/AI-Register/Business Logic/Services/AISystemService.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using BusinessLogic.Classes;
 using BusinessLogic.Entities;
+using BusinessLogic.Enums;
 using BusinessLogic.Interfaces;
 
 namespace BusinessLogic.Services
@@ -25,7 +26,7 @@ namespace BusinessLogic.Services
             foreach (AISystemEntity aiSystem in aiSystemEntities)
             {
                 Provider newProvider = new Provider(aiSystem.ProviderEntity.Id, aiSystem.ProviderEntity.Name, aiSystem.ProviderEntity.Address, aiSystem.ProviderEntity.Email, aiSystem.ProviderEntity.PhoneNumber);
-                AISystem newAISystem = new AISystem(aiSystem.Id, aiSystem.Name, (AIRegisterEnum.AISystemStatus)aiSystem.Status, aiSystem.Description, aiSystem.URL, aiSystem.TechnicalDocumentationLink, (AIRegisterEnum.ApprovalStatus)aiSystem.ApprovalStatus, aiSystem.DateAdded, newProvider, new Certificate(aiSystem.CertificateEntity.Id, aiSystem.CertificateEntity.Type, aiSystem.CertificateEntity.Number, aiSystem.CertificateEntity.ExpiryDate, aiSystem.CertificateEntity.NameNotifiedBody, aiSystem.CertificateEntity.IdNotifiedBody, new ScanCertificate(aiSystem.CertificateEntity.ScanCertificate.Id, aiSystem.CertificateEntity.ScanCertificate.Filename, aiSystem.CertificateEntity.ScanCertificate.Filepath)),(AIRegisterEnum.MemberStates)aiSystem.MemberState);
+                AISystem newAISystem = new AISystem(aiSystem.Id, aiSystem.Name, (AISystemStatus)aiSystem.Status, aiSystem.Description, aiSystem.URL, aiSystem.TechnicalDocumentationLink, (ApprovalStatus)aiSystem.ApprovalStatus, aiSystem.DateAdded, newProvider, new Certificate(aiSystem.CertificateEntity.Id, aiSystem.CertificateEntity.Type, aiSystem.CertificateEntity.Number, aiSystem.CertificateEntity.ExpiryDate, aiSystem.CertificateEntity.NameNotifiedBody, aiSystem.CertificateEntity.IdNotifiedBody, new ScanCertificate(aiSystem.CertificateEntity.ScanCertificate.Id, aiSystem.CertificateEntity.ScanCertificate.Filename, aiSystem.CertificateEntity.ScanCertificate.Filepath)),(MemberStates)aiSystem.MemberState);
                 aiSystems.Add(newAISystem);
             }
             return aiSystems;
@@ -35,7 +36,7 @@ namespace BusinessLogic.Services
         {
             AISystemEntity aiSystem = _IaiSystemRepository.GetAiSystemById(id);
             Provider provider = new Provider(aiSystem.ProviderEntity.Id, aiSystem.ProviderEntity.Name, aiSystem.ProviderEntity.Address, aiSystem.ProviderEntity.Email, aiSystem.ProviderEntity.PhoneNumber);
-            AISystem detailedAiSystem = new AISystem(aiSystem.Id, aiSystem.Name, (AIRegisterEnum.AISystemStatus)aiSystem.Status, aiSystem.Description, aiSystem.URL, aiSystem.TechnicalDocumentationLink, (AIRegisterEnum.ApprovalStatus)aiSystem.ApprovalStatus, aiSystem.DateAdded, provider, new Certificate(aiSystem.CertificateEntity.Id, aiSystem.CertificateEntity.Type, aiSystem.CertificateEntity.Number, aiSystem.CertificateEntity.ExpiryDate, aiSystem.CertificateEntity.NameNotifiedBody, aiSystem.CertificateEntity.IdNotifiedBody, new ScanCertificate(aiSystem.CertificateEntity.ScanCertificate.Id, aiSystem.CertificateEntity.ScanCertificate.Filename, aiSystem.CertificateEntity.ScanCertificate.Filepath)), (AIRegisterEnum.MemberStates)aiSystem.MemberState);
+            AISystem detailedAiSystem = new AISystem(aiSystem.Id, aiSystem.Name, (AISystemStatus)aiSystem.Status, aiSystem.Description, aiSystem.URL, aiSystem.TechnicalDocumentationLink, (ApprovalStatus)aiSystem.ApprovalStatus, aiSystem.DateAdded, provider, new Certificate(aiSystem.CertificateEntity.Id, aiSystem.CertificateEntity.Type, aiSystem.CertificateEntity.Number, aiSystem.CertificateEntity.ExpiryDate, aiSystem.CertificateEntity.NameNotifiedBody, aiSystem.CertificateEntity.IdNotifiedBody, new ScanCertificate(aiSystem.CertificateEntity.ScanCertificate.Id, aiSystem.CertificateEntity.ScanCertificate.Filename, aiSystem.CertificateEntity.ScanCertificate.Filepath)), (MemberStates)aiSystem.MemberState);
             detailedAiSystem.setFiles(aiSystem);
 
             return detailedAiSystem;
@@ -78,7 +79,7 @@ namespace BusinessLogic.Services
         
             AISystemEntity returnAISystem = _IaiSystemRepository.AddSystemAI(aiSystemEntity);
             aiSystem.Guid = returnAISystem.Id;
-            aiSystem.ApprovalStatus = (AIRegisterEnum.ApprovalStatus)returnAISystem.ApprovalStatus;
+            aiSystem.ApprovalStatus = (ApprovalStatus)returnAISystem.ApprovalStatus;
             aiSystem.DateAdded = returnAISystem.DateAdded;
             return aiSystem;
         }
